Add some form of speech recognition
Currently speech recognition is pretty limited: although it can
recognize a few commands at once, web browsers quickly stop speech
recognition. To relaunch it, the user needs to give the authorization
to use the microphone again. A workaround should be found for this.

The other issue is the accessibility of functions to execute when a
command is recognized and execute it. Currently, the functions needed
are unavailable. Maybe moving them to dedicated Javascript files could
help with that.

What the current state of this code does is recognize keywords that
can be in a full sentence ("page suivante" can be recognized in "et
comme on peut le voir dans la page suivante,...") a few times before
speech recognition ends.

Speech recognition will also share its status in the console, in case
we want to see what exactly has been recognized, with which
confidence, or if it failed to recognize anything.